Sync updates: round-2 review fixes, formal quality metrics, coordinate precision & preflight audits

+++ src/qwen_qk_waft/infer.py
@@ -71,7 +71,7 @@ def infer(
             )
 
     native_map = resize_absolute_map(
-        output["final_map"],
+        output["final_map"].float(),
         native_size,
         source_size_from=work_size,
         source_size_to=native_size,
